requirements.txt
django>4,<5
python-decouple
django-filter
markdown
djangorestframework
drf-yasg
djoser
djangorestframework_simplejwt

drf-spectacular

pytest
pytest-django

black